LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-007-expected.txt


FAIL #target3 is scroll-adjusted in both axises assert_equals: expected 100 but got 0
